Update TestNG instrumentations to have separate modules for versions 6 and 7
