+++ dd-java-agent/instrumentation/testng/testng-6/build.gradle
@@ -8,9 +8,8 @@ muzzle {
   pass {
     group = 'org.testng'
     module = 'testng'
-    versions = '[6.4,7.5)'
+    versions = '[6.4,7.0.0)'
     skipVersions += "6.13" // depends on non-existent jcommander-1.66
-    skipVersions += "7.1.0" //  depends on non-existent guice-4.1.0-no_aop
   }
 }
 
@@ -20,10 +19,14 @@ dependencies {
   implementation project(':dd-java-agent:instrumentation:testng')
 
   testImplementation testFixtures(project(':dd-java-agent:instrumentation:testng'))
-
-  testImplementation(group: 'org.testng', name: 'testng') {
-    version {
-      strictly '6.4'
-    }
-  }
+  testImplementation group: 'org.testng', name: 'testng', version: '6.4'
 }
+
+// gradle can't downgrade the testng dependencies with `strictly`
+// and IntelliJ IDEA reports an error when importing the project
+configurations.matching({ it.name.startsWith('test') }).each({
+  it.resolutionStrategy {
+    force group: 'org.testng', name: 'testng', version: '6.4'
+  }
+})
+
