+++ dd-java-agent/instrumentation/testng/testng-6/src/test/groovy/TestNG6Test.groovy
@@ -1,6 +1,6 @@
 import datadog.trace.instrumentation.testng.TestNGTest
 
-class TestNG64Test extends TestNGTest {
+class TestNG6Test extends TestNGTest {
 
   @Override
   String expectedTestFrameworkVersion() {
