+++ dd-java-agent/instrumentation/testng/testng-7/src/test/groovy/TestNG7Test.groovy
@@ -1,16 +1,16 @@
 import datadog.trace.bootstrap.instrumentation.api.Tags
 import datadog.trace.instrumentation.testng.TestNGTest
 
-class TestNG75Test extends TestNGTest {
+class TestNG7Test extends TestNGTest {
 
   @Override
   String expectedTestFrameworkVersion() {
-    return '7.5'
+    return '7.0.0'
   }
 
   @Override
   String assertionErrorMessage() {
-    "expected [true] but found [false]"
+    "did not expect to find [true] but found [false]"
   }
 
   @Override
