+++ settings.gradle
@@ -385,8 +385,8 @@ include ':dd-java-agent:instrumentation:spymemcached-2.10'
 include ':dd-java-agent:instrumentation:sslsocket'
 include ':dd-java-agent:instrumentation:synapse-3'
 include ':dd-java-agent:instrumentation:testng'
-include ':dd-java-agent:instrumentation:testng:testng-6.4'
-include ':dd-java-agent:instrumentation:testng:testng-7.5'
+include ':dd-java-agent:instrumentation:testng:testng-6'
+include ':dd-java-agent:instrumentation:testng:testng-7'
 include ':dd-java-agent:instrumentation:thymeleaf'
 include ':dd-java-agent:instrumentation:tomcat-5.5'
 include ':dd-java-agent:instrumentation:tomcat-5.5-common'
